==> 2_biomarkers/full_driver/1_cn_drivers.ipynb <==
source(paste0(dirname(dirname(getwd())),'/map.r'))
source(paste0(HELP_DIR, "shortcuts.r"))

# 0 - Read and process collected driver data
- Just including all drivers output from Linx here

cn_gene <- 
fread("/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/cn_gene.csv")

purities <- fread("/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/purities.csv")

# 1 - Copy number drivers 
- Copy copy number drivers and save

tmp <- 
cn_gene %>% 
 ij(purities %>% se(sampleId, ploidy), by = "sampleId") %>%
 fi(maxCopyNumber > (3*ploidy)) %>% 
 mu( method = ifelse(minCopyNumber > 3*ploidy, "AMP", "PARTIAL_AMP"))

fwrite(tmp, "/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/drivers_full/amp.txt")

tmp <- 
cn_gene %>% 
 ij( purities %>% se(sampleId, ploidy), by = "sampleId" ) %>%
 fi( minCopyNumber < .5 ) %>% 
 mu( method = "DEL" )  

fwrite(tmp, "/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/drivers_full/del.txt")
