==> 2_biomarkers/full_driver/2_sv_drivers.ipynb <==
source(paste0(dirname(dirname(getwd())),'/map.r'))
source(paste0(HELP_DIR, "shortcuts.r"))

# 0 - Read breakends output from LINX

breakends <- fread( paste0(TMP_DIR, "structural_variants/breakend.csv")) 

#breakends %>% fi(sampleId == "ACTN01020036T", gene == "JAK1")

disruptions <- 
breakends %>% 
 fi(disruptive,undisruptedCopyNumber < .5 , biotype == "protein_coding") %>% 
 tm(sampleId, gene, undisruptedCopyNumber, method = "DISRUPTION")

fwrite(disruptions, "/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/drivers_full/disruption.txt")

# 1 - Fusions

fusions <- 
fread( paste0(TMP_DIR, "structural_variants/fusion.csv")) %>% 
 fi(reported) %>% 
 tm(sampleId, gene = name, method = "FUSION") %>% 
 rw() %>% mu(upstream = strsplit(gene, "_")[[1]][1], downstream = strsplit(gene, "_")[[1]][2]) %>% ug()

fwrite(fusions, "/mnt/petasan_immunocomp/datasets/hartwig/biomarkers/database/drivers_full/fusion.txt")
